backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.6
Pillow==10.4.0
numpy==1.26.4
scikit-learn==1.6.1
xgboost==1.7.6
joblib==1.4.2
python-dotenv==1.0.0
requests==2.28.2
supabase==2.20.0
bcrypt==4.1.2
PyJWT==2.8.0
email-validator==2.0.0
opencv-python==4.8.1.78
scikit-image==0.22.0
